Title: 세 정수의 최솟값, Time: 134ms, Memory: 11MB, Status: Wrong Answer - CodeTree

+++ 240306/세 정수의 최솟값/minimum-of-three-numbers.java
@@ -5,16 +5,18 @@ import java.util.*;
 public class Main {
     public static void main(String[] args) {
     Scanner sc = new Scanner(System.in);
-    int a = sc.nextInt();
-    int b = sc.nextInt();
-    int c = sc.nextInt();
+    double a = sc.nextDouble();
+    double b = sc.nextDouble();
+    double c = sc.nextDouble();
     
+
+    //-100 -99 -98 
     if (a >= b && b >= c) {
-        System.out.print(c);
+        System.out.print((int)c);
     } else if (b >= c && c >= a) {
-        System.out.print(a);
+        System.out.print((int)a);
     } else if (c >= a && a >= b) {
-        System.out.print(b);
+        System.out.print((int)b);
     }
     }
 }
